Work directly from spline
