+++ crest/Assets/Crest/Crest/Scripts/Spline.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4591f4383cf567145886435dd7e6bbad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
