+++ crest/Assets/Crest/Crest/Scripts/Spline/SplinePoint.cs
